+++ testutil/keeper/keeper.go
@@ -105,7 +105,7 @@ func NewTestSetup(t testing.TB) (sdk.Context, TestKeepers, TestMsgServers) {
 		[]Channel{},
 	)
 	launchKeeper.SetMonitoringcKeeper(monitoringConsumerKeeper)
-	claimKeeper := initializer.Claim(paramKeeper, authKeeper, bankKeeper)
+	claimKeeper := initializer.Claim(paramKeeper, authKeeper, distrKeeper, bankKeeper)
 	require.NoError(t, initializer.StateStore.LoadLatestVersion())
 
 	// Create a context using a custom timestamp
@@ -201,7 +201,7 @@ func NewTestSetupWithIBCMocks(
 		channelMock,
 	)
 	launchKeeper.SetMonitoringcKeeper(monitoringConsumerKeeper)
-	claimKeeper := initializer.Claim(paramKeeper, authKeeper, bankKeeper)
+	claimKeeper := initializer.Claim(paramKeeper, authKeeper, distrKeeper, bankKeeper)
 	require.NoError(t, initializer.StateStore.LoadLatestVersion())
 
 	// Create a context using a custom timestamp
