+++ testutil/networksuite/networksuite.go
@@ -2,6 +2,7 @@
 package networksuite
 
 import (
+	sdkmath "cosmossdk.io/math"
 	"math/rand"
 	"strconv"
 
@@ -165,16 +166,19 @@ func populateCampaign(r *rand.Rand, campaignState campaign.GenesisState) campaig
 
 func populateClaim(r *rand.Rand, claimState claim.GenesisState) claim.GenesisState {
 	claimState.AirdropSupply = sample.Coin(r)
-
-	// add claim records
+	totalSupply := sdkmath.ZeroInt()
 	for i := 0; i < 5; i++ {
+		// fill claim records
+		accSupply := sdkmath.NewIntFromUint64(r.Uint64() % 1000)
 		claimRecord := claim.ClaimRecord{
+			Claimable: accSupply,
 			Address:   sample.Address(r),
-			Claimable: sample.Int(r),
 		}
+		totalSupply = totalSupply.Add(accSupply)
 		nullify.Fill(&claimRecord)
 		claimState.ClaimRecords = append(claimState.ClaimRecords, claimRecord)
 	}
+	claimState.AirdropSupply.Amount = totalSupply
 
 	// add missions
 	for i := 0; i < 5; i++ {
