+++ x/profile/keeper/msg_disable_coordinator.go
@@ -5,7 +5,7 @@ import (
 
 	sdkerrors "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	ignterrors "github.com/ignite/modules/errors"
+	ignterrors "github.com/ignite/modules/pkg/errors"
 
 	"github.com/tendermint/spn/x/profile/types"
 )
