+++ x/profile/keeper/msg_update_coordinator_address.go
@@ -6,7 +6,7 @@ import (
 
 	sdkerrors "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	ignterrors "github.com/ignite/modules/errors"
+	ignterrors "github.com/ignite/modules/pkg/errors"
 
 	"github.com/tendermint/spn/x/profile/types"
 )
